+++ src/library/DIPS.Mobile.UI/API/Camera/Preview/iOS/PreviewViewHandler.cs
@@ -1,0 +1,11 @@
+using Microsoft.Maui.Handlers;
+
+namespace DIPS.Mobile.UI.API.Camera.Preview;
+
+internal partial class PreviewViewHandler() : ViewHandler<PreviewView, iOS.PreviewView>(ViewMapper)
+{
+    protected override iOS.PreviewView CreatePlatformView()
+    {
+        return new iOS.PreviewView();
+    }
+}