+++ src/library/DIPS.Mobile.UI/Components/Buttons/Android/ButtonHandler.cs
@@ -82,7 +82,8 @@ public partial class ButtonHandler : Microsoft.Maui.Handlers.ButtonHandler
         // Small delay to wait for initialization of hitrect
         await Task.Delay(1);
 
-        handler.PlatformView.SetAdditionalHitBoxSize(button, button.AdditionalHitBoxSize, handler.MauiContext!);
+        if(button is not null && handler.PlatformView is not null)
+            handler.PlatformView.SetAdditionalHitBoxSize(button, button.AdditionalHitBoxSize, handler.MauiContext!);
     }
 
     public static void MapOverrideCornerRadius(ButtonHandler handler, Button button)